reintroduce webview for other bot config options
