+++ packages/app/ui/components/SettingsScreenView.tsx
@@ -33,6 +33,11 @@ interface Props {
 }
 
 export function SettingsScreenView(props: Props) {
+  const botSettingsFocused =
+    props.focusedRouteName === 'BotSettings' ||
+    props.focusedRouteName === 'BotMcpSettings' ||
+    props.focusedRouteName === 'BotOtherSettings';
+
   const handleLogoutPressed = () => {
     if (isWeb) {
       return;
@@ -111,7 +116,7 @@ export function SettingsScreenView(props: Props) {
               leftIcon="Face"
               rightIcon={'ChevronRight'}
               onPress={props.onBotSettingsPressed}
-              isFocused={props.focusedRouteName === 'BotSettings'}
+              isFocused={botSettingsFocused}
             />
           )}
           <SettingsAction
